requirements.txt
pandas>=2.0.0
lxml>=4.9.0
numpy>=1.24.0
requests>=2.31.0
pandas-ta>=0.3.14b
python-telegram-bot>=20.0
python-dotenv>=1.0.0
tqdm>=4.66.0
tabulate>=0.9.0
matplotlib>=3.7.0
